+++ scripts/delete_post.php
@@ -17,30 +17,31 @@ if(isset($_GET['id']))
 	analyze_user();
 	verify_logged_in();
 
+
 	$id = $_GET['id'];
 
 	$query = "SELECT b_id, photo FROM postings WHERE id=$id";
 	$result = query_db($query);
-	extract($result[0]);
+	if(isset($result[0]))
+	  {
+		extract($result[0]);
 
-	if(!check_admin())
-	  {	
-		if($GLOBALS['b_id'] != $b_id)
+		if(!check_admin())
+		  {	
+			if($GLOBALS['b_id'] != $b_id)
+			  {
+				disconnect_from_db();
+				return;
+			  }
+		  }
+		push_old_post($b_id);
+		$query = "DELETE FROM postings WHERE id=$id";
+		$result = query_db($query);
+		if($photo != "")
 		  {
-			disconnect_from_db();
-			return;
+			unlink('../images/posts/'.$photo);
 		  }
 	  }
-	push_old_post($b_id);
-
-	$query = "DELETE FROM postings WHERE id='$id'";
-	query_db($query);
-
-	if($photo != "")
-	  {
-		unlink('../images/posts/'.$photo);
-	  }
-
 	disconnect_from_db($link);
   }
 ?>
